requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
python-dotenv
sqlalchemy
jinja2
python-jose
PyJWT
pydantic[email]